feat: update close button content description and rename EmptyMediaState file
